[REFACTORING] Advance of the recommender systems w google api
