+++ .gitignore
@@ -3,8 +3,10 @@ elasticsearch-8.15.3
 logstash-8.16.0
 keys.config
 __pycache__
-local_recm_flask
+previous_work
 http_ca.crt
 data_preprocessing.ipynb
 prerocessing.sql
 final_hospital_data.csv
+local_recm_flask_v3
+vae_pretrained_model.pth
