--- er_recommendation/enriched_filtered_df.csv
@@ -1,2 +1,0 @@
-﻿dutyName,dutyAddr,wgs84Lat,wgs84Lon,is_trauma,dutyTel3,hpid,hv10,hv11,hv22,hv25,hv26,hv27,hv28,hv29,hv3,hv30,hv31,hv35,hv36,hv38,hv39,hv41,hv42,hv5,hv7,hv9,hvamyn,hvangioayn,hvcrrtayn,hvctayn,hvec,hvecmoayn,hvgc,hvhypoayn,hvicc,hvidate,hvincuayn,hvmriayn,hvoc,hvoxyayn,hvs01,hvs02,hvs03,hvs04,hvs05,hvs07,hvs14,hvs17,hvs18,hvs19,hvs21,hvs22,hvs23,hvs25,hvs26,hvs27,hvs28,hvs29,hvs30,hvs31,hvs32,hvs33,hvs34,hvs35,hvs38,hvs54,hvs57,hvs58,hvs59,hvventiayn,hvventisoayn,phpid,rnum,travelTime,distance_km,congestion,travel_time_h,travel_time_m,travel_time_s
-국립중앙의료원,"서울특별시 중구 을지로 245, 국립중앙의료원 (을지로6가)",37.56733955813183,127.00579539705472,True,02-2276-2114,A1100052,N,Y,0,0,0,5,3,2,0,1,2,2,0,11,2,6,Y,Y,Y,12,Y,Y,Y,Y,12,Y,84,Y,8,20241231154906,Y,Y,5,N1,15,3,2,3,6,0,20,13,2,0,50,7,2,7,1,1,1,2,56,1,3,5,5,2,202,0,0,0,5,Y,N,A1100052,1,44.202,0.149,알 수 없음,0,0,44
