--- er_recommendation/mapped_columns_info.csv
@@ -1,97 +1,0 @@
-﻿항목명(영문),항목명(국문),항목설명
-hpid,기관코드,기관ID : 동일테이블에 기관명을 포함하고 있음
-phpid,기관코드,구기관코드
-hvidate,입력일시,입력일시
-hvec,일반,일반(응급실일반병상)
-hvoc,[기타] 수술실,[기타] 수술실
-hvncc,[중환자실] 신생아,[중환자실] 신생아
-hvccc,[중환자실] 흉부외과,[중환자실] 흉부외과
-hvicc,[중환자실] 일반,[중환자실] 일반
-hvgc,[입원실] 일반,[입원실] 일반
-hvctayn,CT가용(가/부),CT가용(가/부)
-hvmriayn,MRI가용(가/부),MRI가용(가/부)
-hvangioayn,혈관촬영기가용(가/부),혈관촬영기가용(가/부)
-hvventiayn,인공호흡기가용(가/부),인공호흡기가용(가/부)
-hvamyn,구급차가용여부,구급차가용여부
-hv2,[중환자실] 내과,[중환자실] 내과
-hv3,[중환자실] 외과,[중환자실] 외과
-hv5,신경과입원실,신경과입원실
-hv6,[중환자실] 신경외과,[중환자실] 신경외과
-hv7,약물중환자,약물중환자
-hv10,VENTI(소아),VENTI(소아)
-hv11,인큐베이터(보육기),인큐베이터(보육기)
-dutyName,,
-dutyTel3,,
-hv28,소아,소아
-hv29,응급실 음압 격리 병상,응급실 음압 격리 병상
-hv30,응급실 일반 격리 병상,응급실 일반 격리 병상
-hv34,[중환자실] 심장내과,[중환자실] 심장내과
-hv41,[입원실] 음압격리,[입원실] 음압격리
-hv42,[기타] 분만실,[기타] 분만실
-hvcrrtayn,CRRT가용(가/부),CRRT가용(가/부)
-hvecmoayn,ECMO가용(가/부),ECMO가용(가/부)
-hvhypoayn,중심체온조절유도기(가/부),중심체온조절유도기
-hvincuayn,인큐베이터가용(가/부),인큐베이터가용(가/부)
-hvoxyayn,고압산소치료기가용(가/부),고압산소치료기가용(가/부)
-hvs01,,
-hvs02,,
-hvs03,,
-hvs04,,
-hvs08,,
-hvs12,,
-hvs15,,
-hvs17,,
-hvs22,,
-hvs25,,
-hvs26,,
-hvs27,,
-hvs28,,
-hvs29,,
-hvs30,,
-hvs31,,
-hvs32,,
-hvs33,,
-hvs34,,
-hvs35,,
-hvs38,,
-hvventisoayn,인공호흡기 조산아가용(가/부),인공호흡기 조산아가용(가/부)
-rnum,,
-hv32,[중환자실] 소아,[중환자실] 소아
-hv35,[중환자실] 음압격리,[중환자실] 음압격리
-hv40,[입원실] 정신과 폐쇄병동,[입원실] 정신과 폐쇄병동
-hvs06,,
-hvs07,,
-hvs09,,
-hvs16,,
-hvs18,,
-hvs24,,
-hvcc,[중환자실] 신경과,[중환자실] 신경과
-hvdnm,당직의,당직의
-hv1,응급실 당직의 직통연락처,응급실 당직의 직통연락처
-hv4,외과입원실(정형외과),외과입원실(정형외과)
-hv8,[중환자실] 화상,[중환자실] 화상
-hv9,[중환자실] 외상,[중환자실] 외상
-hv12,소아당직의 직통연락처,소아당직의 직통연락처
-hv13,격리진료구역 음압격리병상,격리진료구역 음압격리병상
-hv14,격리진료구역 일반격리병상,격리진료구역 일반격리병상
-hv15,소아음압격리,소아음압격리
-hv16,소아일반격리,소아일반격리
-hv17,[응급전용] 중환자실 음압격리,[응급전용] 중환자실 음압격리
-hv18,[응급전용] 중환자실 일반격리,[응급전용] 중환자실 일반격리
-hv19,[응급전용] 입원실 음압격리,[응급전용] 입원실 음압격리
-hv21,[응급전용] 입원실 일반격리,[응급전용] 입원실 일반격리
-hv22,감염병 전담병상 중환자실,감염병 전담병상 중환자실
-hv23,감염병 전담병상 중환자실 내 음압격리병상,감염병 전담병상 중환자실 내 음압격리병상
-hv24,[감염] 중증 병상,[감염] 중증 병상
-hv25,[감염] 준-중증 병상,[감염] 준-중증 병상
-hv26,[감염] 중등증 병상,[감염] 중등증 병상
-hv27,코호트 격리,코호트 격리
-hv31,[응급전용] 중환자실,[응급전용] 중환자실
-hv33,[응급전용] 소아중환자실,[응급전용] 소아중환자실
-hv36,[응급전용] 입원실,[응급전용] 입원실
-hv37,[응급전용] 소아입원실,[응급전용] 소아입원실
-hv38,[입원실] 외상전용,[입원실] 외상전용
-hv39,[기타] 외상전용 수술실,[기타] 외상전용 수술실
-hv43,[기타] 화상전용처치실,[기타] 화상전용처치실
-dutyname,기관명,기관명
-dutytel3,응급실전화,응급실전화
