--- hospital_recommendation/utils/basic_autoencoder.py
@@ -1,22 +1,0 @@
-import torch.nn as nn
-
-class AutoEncoder(nn.Module):
-    def __init__(self, input_dim, latent_dim):
-        super(AutoEncoder, self).__init__()
-        self.encoder = nn.Sequential(
-            nn.Linear(input_dim, 128),
-            nn.LeakyReLU(),
-            nn.Linear(128, latent_dim),
-            nn.LeakyReLU()
-        )
-        self.decoder = nn.Sequential(
-            nn.Linear(latent_dim, 128),
-            nn.LeakyReLU(),
-            nn.Linear(128, input_dim),
-            nn.Sigmoid()
-        )
-
-    def forward(self, x):
-        latent = self.encoder(x)
-        reconstructed = self.decoder(latent)
-        return latent, reconstructed
