--- recm_flask/requirements.txt
@@ -1,9 +1,0 @@
-flask==3.0.0
-elasticsearch==8.17.0
-sentence-transformers==3.3.1
-torch==2.5.1
-numpy==1.26.0
-scikit-learn==1.4.0
-redis==5.2.1
-pandas==2.2.3
-openai
